smart-meal-planner/requirements.txt
fastapi>=0.111.0
uvicorn>=0.29.0
pydantic>=2.7.1
langchain>=0.2.1
langchain-openai>=0.1.8
langchain-community>=0.2.1
faiss-cpu>=1.12.0
sentence-transformers>=3.0.0
openai>=1.30.1
sqlalchemy>=2.0.30
streamlit>=1.35.0
requests>=2.32.2
python-dotenv>=1.0.1
tiktoken>=0.7.0
numpy>=1.26.4
pandas>=2.2.2
httpx>=0.27.0
